philo

+++ philo/includes/philo.h
@@ -12,10 +12,10 @@ typedef struct s_philo
 	int				id;
 	int				eating;
 	int				meals_eaten;
-	size_t			last_meal;
-	size_t			death_time;
-	size_t			eat_time;
-	size_t			sleep_time;
+	long long		last_meal;
+	long long		death_time;
+	long long		eat_time;
+	long long		sleep_time;
 	long long		start_time;
 	int				nb_philos;
 	int				nb_to_eat;
@@ -46,18 +46,17 @@ void		init_program(t_program *program, t_philo *philos);
 void 		init_forks(pthread_mutex_t *forks, int philo_num);
 void 		edge_philo(char **argv, t_philo *philos, int i, pthread_mutex_t *forks);
 void		init_philos(t_philo *philos, t_program *program, pthread_mutex_t *forks, char **argv);
-void init_threads(t_program *program, pthread_mutex_t *forks);
-void destroy_mutex(int i, t_program *program, pthread_mutex_t *forks);
+void		init_threads(t_program *program, pthread_mutex_t *forks);
+void		destroy_mutex(int i, t_program *program, pthread_mutex_t *forks);
 void		*philosopher_routine(void *ptr);
-int			philosopher_dead(t_philo *philo, size_t time_to_die);
+int			philosopher_dead(t_philo *philo, long long time_to_die);
 void		print_message(char c, t_philo *philo);
 void 		ft_sleep(t_philo *philo);
 void		think(t_philo *philo);
-void eat(t_philo *philo);
-void get_that_fork(t_philo *philo);
+void		eat(t_philo *philo);
+void		get_that_fork(t_philo *philo);
 int			is_he_dead(t_philo *philo);
-void one_philo(int death_time);
-
+void		one_philo(int death_time);
 
 enum {
 	KO = 1,
